Fixed Grammatical Errors in Community Theme Descriptions (#1376)
* Fixed grammatical error in theme-data.js

In the Flatlogic description, the phrase "how real applications is built" is grammatically incorrect since "applications" is a plural noun but "is" refers to a singular noun. I changed it to "how real applications are built."

In the PrimeVue description, the phrase "look&feel" (which lacked spaces between the words and punctuation) has been changed to "look & feel".

* Update theme-data.js

In the Flatlogic description, "Check the admin dashboard templates" has been changed to "Check out the admin dashboard templates."